+++ src/bookmark_reader/firefox.rs
@@ -1,5 +1,8 @@
 use super::{ReadBookmark, SelectSource, SourceOs};
-use crate::{bookmarks::SourceBookmarkBuilder, utils, Source, SourceBookmarks, SourceType};
+use crate::{
+    bookmark_reader::json_reader::traverse_json, bookmarks::SourceBookmarkBuilder, utils, Source,
+    SourceBookmarks, SourceType,
+};
 use anyhow::anyhow;
 use log::{debug, trace};
 use serde_json::{Map, Value};
@@ -153,7 +156,12 @@ impl FirefoxReader {
         Box::new(Self)
     }
 
-    fn select_bookmark(obj: &Map<String, Value>, source: &Source, bookmarks: &mut SourceBookmarks) {
+    fn select_bookmark(
+        obj: &Map<String, Value>,
+        source: &Source,
+        bookmarks: &mut SourceBookmarks,
+        folder: &mut Option<String>,
+    ) {
         trace!("json object: {obj:#?}");
 
         if let Some(Value::String(type_value)) = obj.get("type") {
@@ -161,7 +169,8 @@ impl FirefoxReader {
                 if let Some(Value::String(uri_value)) = obj.get("uri") {
                     if uri_value.contains("http") {
                         let source_bookmark = SourceBookmarkBuilder::new(uri_value)
-                            .add_source(&source.source_type)
+                            .add_source(source.source_type.to_owned())
+                            .add_folder_opt(source.source_type.to_owned(), folder.to_owned())
                             .build();
                         bookmarks.insert(source_bookmark);
                     }
@@ -170,64 +179,16 @@ impl FirefoxReader {
         }
     }
 
-    pub fn traverse_json(value: &Value, source: &Source, bookmarks: &mut SourceBookmarks) {
-        match value {
-            Value::Object(obj) => {
-                if source.folders.is_empty() {
-                    Self::select_bookmark(obj, source, bookmarks);
-
-                    for (_, val) in obj {
-                        Self::traverse_json(val, source, bookmarks);
-                    }
-                } else {
-                    if let Some(Value::String(type_value)) = obj.get("type") {
-                        if type_value == "text/x-moz-place-container" {
-                            if let Some(Value::String(title_value)) = obj.get("title") {
-                                if source.folders.contains(title_value) {
-                                    for (_, val) in obj {
-                                        Self::traverse_children(val, source, bookmarks);
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    for (_, val) in obj {
-                        Self::traverse_json(val, source, bookmarks);
-                    }
+    fn select_folder(obj: &Map<String, Value>) -> Option<&String> {
+        if let Some(Value::String(type_value)) = obj.get("type") {
+            if type_value == "text/x-moz-place-container" {
+                if let Some(Value::String(title)) = obj.get("title") {
+                    return Some(title);
                 }
             }
-            Value::Array(arr) => {
-                for val in arr {
-                    Self::traverse_json(val, source, bookmarks);
-                }
-            }
-            Value::String(_) => (),
-            Value::Number(_) => (),
-            Value::Bool(_) => (),
-            Value::Null => (),
         }
-    }
 
-    fn traverse_children(value: &Value, source: &Source, bookmarks: &mut SourceBookmarks) {
-        match value {
-            Value::Object(obj) => {
-                Self::select_bookmark(obj, source, bookmarks);
-
-                for (_, val) in obj {
-                    Self::traverse_children(val, source, bookmarks);
-                }
-            }
-            Value::Array(arr) => {
-                for val in arr {
-                    Self::traverse_children(val, source, bookmarks);
-                }
-            }
-            Value::String(_) => (),
-            Value::Number(_) => (),
-            Value::Bool(_) => (),
-            Value::Null => (),
-        }
+        None
     }
 }
 
@@ -270,7 +231,13 @@ impl<'a> ReadBookmark<'a> for FirefoxReader {
         source_bookmarks: &mut SourceBookmarks,
     ) -> Result<(), anyhow::Error> {
         debug!("Import bookmarks from {:#?}", self.name());
-        Self::traverse_json(&parsed_bookmarks, source, source_bookmarks);
+        traverse_json(
+            &parsed_bookmarks,
+            source,
+            source_bookmarks,
+            Self::select_bookmark,
+            Self::select_folder,
+        );
         Ok(())
     }
 }
@@ -280,8 +247,7 @@ mod tests {
     use super::*;
     use crate::{
         bookmark_reader::{
-            source_reader::{CompressedJsonReader, JsonReader},
-            ParsedBookmarks, ReadSource, SourceReader,
+            CompressedJsonReader, JsonReader, ParsedBookmarks, ReadSource, SourceReader,
         },
         test_utils, utils,
     };
@@ -501,25 +467,25 @@ mod tests {
                 (
                     url1.to_owned(),
                     SourceBookmarkBuilder::new(url1)
-                        .add_source(&SourceType::Firefox)
+                        .add_source(SourceType::Firefox)
                         .build()
                 ),
                 (
                     url2.to_owned(),
                     SourceBookmarkBuilder::new(url2)
-                        .add_source(&SourceType::Firefox)
+                        .add_source(SourceType::Firefox)
                         .build()
                 ),
                 (
                     url3.to_owned(),
                     SourceBookmarkBuilder::new(url3)
-                        .add_source(&SourceType::Firefox)
+                        .add_source(SourceType::Firefox)
                         .build()
                 ),
                 (
                     url4.to_owned(),
                     SourceBookmarkBuilder::new(url4)
-                        .add_source(&SourceType::Firefox)
+                        .add_source(SourceType::Firefox)
                         .build()
                 )
             ])
@@ -552,13 +518,15 @@ mod tests {
                 (
                     url1.to_owned(),
                     SourceBookmarkBuilder::new(url1)
-                        .add_source(&SourceType::Firefox)
+                        .add_source(SourceType::Firefox)
+                        .add_folder(SourceType::Firefox, "dev")
                         .build()
                 ),
                 (
                     url2.to_owned(),
                     SourceBookmarkBuilder::new(url2)
-                        .add_source(&SourceType::Firefox)
+                        .add_source(SourceType::Firefox)
+                        .add_folder(SourceType::Firefox, "rust")
                         .build()
                 ),
             ])